Fixes a race condition when Livewire uploads is used on busy websites (#2492)
* Fixes a race condition when Livewire uploads is used on a high-traffic website

* Incorporates a suggestion from @nuernbergerA to reduce cyclomatic complexity

+++ src/WithFileUploads.php
@@ -96,6 +96,10 @@ trait WithFileUploads
         $storage = FileUploadConfiguration::storage();
 
         foreach ($storage->allFiles(FileUploadConfiguration::path()) as $filePathname) {
+            if (! $storage->exists($filePathname)) {
+                continue;
+            }
+
             $yesterdaysStamp = now()->subDay()->timestamp;
             if ($yesterdaysStamp > $storage->lastModified($filePathname)) {
                 $storage->delete($filePathname);
